for wait version up AWS SDK
