--- aws_paginator.go
@@ -1,147 +1,0 @@
-package stefunny
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/aws/aws-sdk-go-v2/service/eventbridge"
-	"github.com/aws/aws-sdk-go-v2/service/sfn"
-)
-
-type ListStateMachineAliasesPaginator struct {
-	client    SFnClient
-	params    *sfn.ListStateMachineAliasesInput
-	nextToken *string
-	firstPage bool
-}
-
-func NewListStateMachineAliasesPaginator(client SFnClient, params *sfn.ListStateMachineAliasesInput) *ListStateMachineAliasesPaginator {
-	if params == nil {
-		params = &sfn.ListStateMachineAliasesInput{}
-	}
-
-	return &ListStateMachineAliasesPaginator{
-		client:    client,
-		params:    params,
-		firstPage: true,
-	}
-}
-
-func (p *ListStateMachineAliasesPaginator) HasMorePages() bool {
-	return p.firstPage || p.nextToken != nil
-}
-
-func (p *ListStateMachineAliasesPaginator) NextPage(ctx context.Context, optFns ...func(*sfn.Options)) (*sfn.ListStateMachineAliasesOutput, error) {
-	if !p.HasMorePages() {
-		return nil, fmt.Errorf("no more pages available")
-	}
-
-	params := *p.params
-	params.NextToken = p.nextToken
-
-	result, err := p.client.ListStateMachineAliases(ctx, &params, optFns...)
-	if err != nil {
-		return nil, err
-	}
-	p.firstPage = false
-
-	prevToken := p.nextToken
-	p.nextToken = result.NextToken
-
-	if prevToken != nil && p.nextToken != nil && *prevToken == *p.nextToken {
-		p.nextToken = nil
-	}
-	return result, nil
-}
-
-type ListStateMachineVersionsPaginator struct {
-	client    SFnClient
-	params    *sfn.ListStateMachineVersionsInput
-	nextToken *string
-	firstPage bool
-}
-
-func NewListStateMachineVersionsPaginator(client SFnClient, params *sfn.ListStateMachineVersionsInput) *ListStateMachineVersionsPaginator {
-	if params == nil {
-		params = &sfn.ListStateMachineVersionsInput{}
-	}
-
-	return &ListStateMachineVersionsPaginator{
-		client:    client,
-		params:    params,
-		firstPage: true,
-	}
-}
-
-func (p *ListStateMachineVersionsPaginator) HasMorePages() bool {
-	return p.firstPage || p.nextToken != nil
-}
-
-func (p *ListStateMachineVersionsPaginator) NextPage(ctx context.Context, optFns ...func(*sfn.Options)) (*sfn.ListStateMachineVersionsOutput, error) {
-	if !p.HasMorePages() {
-		return nil, fmt.Errorf("no more pages available")
-	}
-
-	params := *p.params
-	params.NextToken = p.nextToken
-
-	result, err := p.client.ListStateMachineVersions(ctx, &params, optFns...)
-	if err != nil {
-		return nil, err
-	}
-	p.firstPage = false
-
-	prevToken := p.nextToken
-	p.nextToken = result.NextToken
-
-	if prevToken != nil && p.nextToken != nil && *prevToken == *p.nextToken {
-		p.nextToken = nil
-	}
-	return result, nil
-}
-
-type ListRuleNamesByTargetPaginator struct {
-	client    EventBridgeClient
-	params    *eventbridge.ListRuleNamesByTargetInput
-	nextToken *string
-	firstPage bool
-}
-
-func NewListRuleNamesByTargetPaginator(client EventBridgeClient, params *eventbridge.ListRuleNamesByTargetInput) *ListRuleNamesByTargetPaginator {
-	if params == nil {
-		params = &eventbridge.ListRuleNamesByTargetInput{}
-	}
-
-	return &ListRuleNamesByTargetPaginator{
-		client:    client,
-		params:    params,
-		firstPage: true,
-	}
-}
-
-func (p *ListRuleNamesByTargetPaginator) HasMorePages() bool {
-	return p.firstPage || p.nextToken != nil
-}
-
-func (p *ListRuleNamesByTargetPaginator) NextPage(ctx context.Context, optFns ...func(*eventbridge.Options)) (*eventbridge.ListRuleNamesByTargetOutput, error) {
-	if !p.HasMorePages() {
-		return nil, fmt.Errorf("no more pages available")
-	}
-
-	params := *p.params
-	params.NextToken = p.nextToken
-
-	result, err := p.client.ListRuleNamesByTarget(ctx, &params, optFns...)
-	if err != nil {
-		return nil, err
-	}
-	p.firstPage = false
-
-	prevToken := p.nextToken
-	p.nextToken = result.NextToken
-
-	if prevToken != nil && p.nextToken != nil && *prevToken == *p.nextToken {
-		p.nextToken = nil
-	}
-	return result, nil
-}
